Refactor CommittedStatsStore / Enricher slightly. Enable CommittedStatsStore / Enricher in both socket and STDIO modes. Move packages and delete interfaces accordingly.
