--- airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/dataflow/state/stats/CommittedStatsStore.kt
@@ -1,17 +1,0 @@
-/*
- * Copyright (c) 2025 Airbyte, Inc., all rights reserved.
- */
-
-package io.airbyte.cdk.load.dataflow.state.stats
-
-import io.airbyte.cdk.load.command.DestinationStream
-import io.airbyte.cdk.load.dataflow.state.PartitionHistogram
-
-/** tallies "committed" record and byte counts from histograms on aggregate flush */
-interface CommittedStatsStore {
-    fun acceptStats(
-        s: DestinationStream.Descriptor,
-        counts: PartitionHistogram,
-        bytes: PartitionHistogram,
-    )
-}
